Revert "ACMS-138: Use trait-based approach"
This reverts commit 519122b3c273a63c2f1e1ae6ae553c4c6da05355.
