+++ modules/acquia_cms_common/tests/src/FunctionalJavascript/MediaEmbedTestBase.php
@@ -3,11 +3,12 @@
 namespace Drupal\Tests\acquia_cms_common\FunctionalJavascript;
 
 use Acquia\DrupalEnvironmentDetector\AcquiaDrupalEnvironmentDetector;
+use Behat\Mink\Element\ElementInterface;
 use Drupal\FunctionalJavascriptTests\WebDriverTestBase;
 use Drupal\media\Entity\MediaType;
-use Drupal\Tests\acquia_cms_common\Traits\MediaLibraryCreationTrait;
 use Drupal\Tests\acquia_cms_common\Traits\MediaTestTrait;
 use Drupal\Tests\ckeditor\Traits\CKEditorTestTrait;
+use Drupal\Tests\TestFileCreationTrait;
 
 /**
  * Base class for testing CKEditor embeds of a specific media type.
@@ -16,6 +17,7 @@ abstract class MediaEmbedTestBase extends WebDriverTestBase {
 
   use CKEditorTestTrait;
   use MediaTestTrait;
+  use TestFileCreationTrait;
 
   /**
    * {@inheritdoc}
@@ -84,17 +86,58 @@ abstract class MediaEmbedTestBase extends WebDriverTestBase {
     $this->selectMedia(0);
     $this->insertSelectedMedia();
     $this->assertMediaIsEmbedded();
-    // Exit the CKEditor iFrame.
-    $this->getSession()->switchToIFrame(NULL);
+  }
 
-    // If this class wants to test creating media in the media library, do it.
-    if (in_array(MediaLibraryCreationTrait::class, class_uses(static::class), TRUE)) {
-      $this->openMediaLibrary();
-      $this->addMedia();
-      $added_media = $this->assertSession()->waitForElementVisible('css', '.js-media-library-add-form-added-media > li');
-      $this->assertNotEmpty($added_media);
-      $this->assertAddedMedia($added_media);
-    }
+  /**
+   * Returns the path of a test file to upload into the media library.
+   *
+   * @param string $type
+   *   The file type. This can be any of the types accepted by
+   *   \Drupal\Tests\TestFileCreationTrait::getTestFiles().
+   *
+   * @return string
+   *   The absolute path to the test file.
+   */
+  protected function getTestFilePath($type) {
+    $files = $this->getTestFiles($type);
+    $this->assertNotEmpty($files);
+    $uri = reset($files)->uri;
+    $path = $this->container->get('file_system')->realpath($uri);
+    $this->assertNotEmpty($path);
+    $this->assertFileExists($path);
+    return $path;
+  }
+
+  /**
+   * Tests creating a new media item in the media library.
+   */
+  protected function doTestCreateMedia() {
+    $this->openMediaLibrary();
+    $this->addMedia();
+    $added_media = $this->assertSession()->waitForElementVisible('css', '.js-media-library-add-form-added-media > li');
+    $this->assertNotEmpty($added_media);
+    $this->assertAddedMedia($added_media);
+  }
+
+  /**
+   * Asserts required fields of a media item being created in the media library.
+   *
+   * @param \Behat\Mink\Element\ElementInterface $added_media
+   *   The element containing the required fields of the media item being
+   *   created.
+   */
+  protected function assertAddedMedia(ElementInterface $added_media) {
+    // Nothing to do by default.
+  }
+
+  /**
+   * Begins creating a media item in the media library.
+   *
+   * Normally this should enter the source field value for the new media item
+   * (i.e., upload an image or file, enter a video URL, etc.)
+   */
+  protected function addMedia() {
+    // Nothing to do by default.
   }
 
   /**
@@ -134,6 +177,9 @@ abstract class MediaEmbedTestBase extends WebDriverTestBase {
    * Opens the media library in CKEditor.
    */
   protected function openMediaLibrary() {
+    // Exit the CKEditor iFrame if we're in it.
+    $this->getSession()->switchToIFrame(NULL);
+
     $this->waitForEditor();
     $this->pressEditorButton('drupalmedialibrary');
