+++ modules/acquia_cms_document/tests/src/FunctionalJavascript/DocumentEmbedTest.php
@@ -4,7 +4,6 @@ namespace Drupal\Tests\acquia_cms_document\Functional;
 
 use Behat\Mink\Element\ElementInterface;
 use Drupal\Tests\acquia_cms_common\FunctionalJavascript\MediaEmbedTestBase;
-use Drupal\Tests\acquia_cms_common\Traits\MediaLibraryCreationTrait;
 
 /**
  * Tests embedding Document media in CKEditor.
@@ -13,8 +12,6 @@ use Drupal\Tests\acquia_cms_common\Traits\MediaLibraryCreationTrait;
  * @group acquia_cms_document
  */
 class DocumentEmbedTest extends MediaEmbedTestBase {
-
-  use MediaLibraryCreationTrait;
 
   /**
    * {@inheritdoc}
@@ -40,6 +37,14 @@ class DocumentEmbedTest extends MediaEmbedTestBase {
    * {@inheritdoc}
    */
   protected $mediaType = 'document';
+
+  /**
+   * {@inheritdoc}
+   */
+  public function testEmbedMedia() {
+    parent::testEmbedMedia();
+    $this->doTestCreateMedia();
+  }
 
   /**
    * {@inheritdoc}
