+++ modules/acquia_cms_image/tests/src/FunctionalJavascript/ImageEmbedTest.php
@@ -4,7 +4,6 @@ namespace Drupal\Tests\acquia_cms_image\Functional;
 
 use Behat\Mink\Element\ElementInterface;
 use Drupal\Tests\acquia_cms_common\FunctionalJavascript\MediaEmbedTestBase;
-use Drupal\Tests\acquia_cms_common\Traits\MediaLibraryCreationTrait;
 
 /**
  * Tests embedding Image media in CKEditor.
@@ -13,8 +12,6 @@ use Drupal\Tests\acquia_cms_common\Traits\MediaLibraryCreationTrait;
  * @group acquia_cms_image
  */
 class ImageEmbedTest extends MediaEmbedTestBase {
-
-  use MediaLibraryCreationTrait;
 
   /**
    * {@inheritdoc}
@@ -59,6 +56,14 @@ class ImageEmbedTest extends MediaEmbedTestBase {
         ],
       ])
       ->save();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function testEmbedMedia() {
+    parent::testEmbedMedia();
+    $this->doTestCreateMedia();
   }
 
   /**
